added first version of performAction and testTrigger into Aunoa-Service

+++ app/src/main/java/de/mseprojekt/aunoa/feature_app/domain/model/triggerObjects/TimeTrigger.kt
@@ -3,8 +3,8 @@ package de.mseprojekt.aunoa.feature_app.domain.model.triggerObjects
 import java.time.DayOfWeek
 
 data class TimeTrigger(
-    val startTime : Long,
-    val endTime : Long,
+    val startTime : Int,
+    val endTime : Int,
     val startWeekday : DayOfWeek?,
     val endWeekday: DayOfWeek?
 ): TriggerObject()
